removePublicKey tests

+++ graphql/resolvers.go
@@ -1406,6 +1406,9 @@ func (r *mutationResolver) CreatePublicKey(ctx context.Context, publicKeyInput P
 }
 
 func (r *mutationResolver) RemovePublicKey(ctx context.Context, keyName string) ([]*restModel.APIPubKey, error) {
+	if !doesPublicKeyNameAlreadyExist(ctx, keyName) {
+		return nil, InputValidationError.Send(ctx, fmt.Sprintf("Error deleting public key. Provided key name, %s, does not exist.", keyName))
+	}
 	err := route.MustHaveUser(ctx).DeletePublicKey(keyName)
 	if err != nil {
 		return nil, InternalServerError.Send(ctx, fmt.Sprintf("Error deleting public key: %s", err.Error()))
